legacy_python/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.7.0
langchain>=0.3.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
gTTS>=2.5.0
google-cloud-texttospeech>=2.15.0
playsound==1.3.0
tqdm>=4.66.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.9
colorama>=0.4.6
fastapi>=0.110.0
uvicorn[standard]>=0.27.0